+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Added
  - Alternative constructors for ``Konfig``
  - Add configuration extending. `#34`_
  - Add casting for container types. `#35`_
+ - Add ``userpass`` authentication method for vault `#18`_
 
 0.6.0 - 2019-06-30
 ------------------
@@ -24,3 +25,4 @@ Added
 .. _#35: https://github.com/kiwicom/konfetti/issues/35
 .. _#34: https://github.com/kiwicom/konfetti/issues/34
 .. _#12: https://github.com/kiwicom/konfetti/issues/12
+.. _#18: https://github.com/kiwicom/konfetti/issues/18